changed data type of remotePort to uint16_t

+++ musicwand_esp32/classes/osc_endpoint.hpp
@@ -12,7 +12,7 @@ class OSCEndpoint {
 
 public:
   OSCEndpoint(WiFiUDP& socket, uint16_t remotePort) : OSCEndpoint(socket, IPAddress(255, 255, 255, 255), remotePort) {}
-  OSCEndpoint(WiFiUDP& socket, IPAddress remoteAddress, uint8_t remotePort) : socket(socket), remoteAddress(remoteAddress), remotePort(remotePort) {}
+  OSCEndpoint(WiFiUDP& socket, IPAddress remoteAddress, uint16_t remotePort) : socket(socket), remoteAddress(remoteAddress), remotePort(remotePort) {}
 
   void sendMessage(const char route[], int32_t value);
 };
